# firewall/openapi/telco-openapi-one.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Insecure Telco Vulnerable API",
    "version": "1.0.0",
    "description": "An intentionally vulnerable API inspired by TMF 622, designed to eventually showcase API security issues like BOLA,BOPLA, BFLA, and other OWASP API Top 10 vulnerabilities."
  },
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/productOrder": {
      "post": {
        "summary": "Create Product Order",
        "operationId": "CreateProductOrder",
        "description": "Creates a new product order based on a product offering.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "$ref": "#/components/schemas/productId"
                  },
                  "quantity": {
                    "$ref": "#/components/schemas/quantity"
                  },
                  "customerId": {
                    "$ref": "#/components/schemas/customerId"
                  }
                },
                "required": [
                  "productId",
                  "quantity",
                  "customerId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product order created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "$ref": "#/components/schemas/orderId"
                    },
                    "message": {
                      "type": "string",
                      "example": "Order created successfully.",
                      "pattern": "^Order created successfully.$",
                      "minLength": 0,
                      "maxLength": 100
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      },
      "get": {
        "summary": "List Product Orders",
        "operationId": "ListProductOrders",
        "description": "Retrieves a list of all product orders.",
        "parameters": [
          {
            "$ref": "#/components/parameters/customerId"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter orders by status.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of product orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "orderId": {
                        "$ref": "#/components/schemas/orderId"
                      },
                      "productId": {
                        "$ref": "#/components/schemas/productId"
                      },
                      "quantity": {
                        "$ref": "#/components/schemas/quantity"
                      },
                      "status": {
                        "$ref": "#/components/schemas/status"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/productOrder/{orderId}": {
      "get": {
        "summary": "Retrieve Product Order",
        "operationId": "RetrieveProductOrder",
        "description": "Fetches details of a specific product order by ID.",
        "parameters": [
          {
            "$ref": "#/components/parameters/orderId"
          }
        ],
        "responses": {
          "200": {
            "description": "Product order details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "$ref": "#/components/schemas/orderId"
                    },
                    "productId": {
                      "$ref": "#/components/schemas/productId"
                    },
                    "quantity": {
                      "$ref": "#/components/schemas/quantity"
                    },
                    "status": {
                      "$ref": "#/components/schemas/status"
                    },
                    "orderDate": {
                      "$ref": "#/components/schemas/orderDate"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      },
      "patch": {
        "summary": "Update Product Order",
        "operationId": "UpdateProductOrder",
        "description": "Partially updates an existing product order.",
        "parameters": [
          {
            "$ref": "#/components/parameters/orderId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "$ref": "#/components/schemas/quantity"
                  },
                  "productId": {
                    "$ref": "#/components/schemas/productId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product order updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "order": {
                      "type": "object",
                      "properties": {
                        "orderId": {
                          "$ref": "#/components/schemas/orderId"
                        },
                        "quantity": {
                          "$ref": "#/components/schemas/quantity"
                        },
                        "priority": {
                          "$ref": "#/components/schemas/priority"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      },
      "delete": {
        "summary": "Cancel Product Order",
        "operationId": "CancelProductOrder",
        "description": "Cancels an existing product order by ID.",
        "parameters": [
          {
            "$ref": "#/components/parameters/orderId"
          }
        ],
        "responses": {
          "200": {
            "description": "Product order canceled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "orderId": {
        "name": "orderId",
        "description": "The ID of the order to retrieve.",
        "required": true,
        "in": "path",
        "schema": {
          "$ref": "#/components/schemas/orderId"
        }
      },
      "customerId": {
        "name": "customerId",
        "description": "The ID of the customer to filter orders.",
        "required": false,
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/customerId"
        }
      },
...
    },
    "schemas": {
...
      "productId": {
        "type": "string"
      },
      "customerId": {
        "type": "string"
      },
      "priority": {
        "type": "string"
      },
      "quantity": {
        "type": "integer"
      },
      "status": {
        "type": "string"
      },
      "orderId": {
        "type": "string"
      },
...
      "orderDate": {
        "type": "string"
      },
      "MessageObject": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ResponseMessage"
          }
        }
      },
      "ResponseMessage": {
        "type": "string"
      }
    },
    "responses": {
      "DefaultErrorResponse": {
        "description": "Unexpected Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError400": {
        "description": "Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
...
      "ResponseError406": {
        "description": "Not Acceptable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError415": {
        "description": "Unsupported Media Type",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError429": {
        "description": "Too Many Requests",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      }
    }
  }
}